genai-story-generator-UlageshS-2000147168/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-dotenv==1.0.1
openai==1.51.0
azure-ai-openai==0.3.0
